requirements/tests.txt
# Additional dependencies for the test suite

# Indirect dependencies first, exact versions for consistency

# ipdb -> ipython -> pexpect
ptyprocess==0.5.1; sys_platform != 'win32' and platform_python_implementation != 'PyPy'

# ipdb -> ipython -> pickleshare
path.py==9.0; platform_python_implementation != 'PyPy'

# ipdb -> ipython -> prompt_toolkit, ipdb -> ipython -> traitlets
six==1.10.0; platform_python_implementation != 'PyPy'

# ipdb -> ipython -> prompt_toolkit
wcwidth==0.1.7; platform_python_implementation != 'PyPy'

# ipdb -> ipython -> traitlets
decorator==4.0.10; platform_python_implementation != 'PyPy'
enum34==1.1.6; platform_python_implementation != 'PyPy' and python_version == '2.7'
ipython_genutils==0.1.0; platform_python_implementation != 'PyPy'

# ipdb -> ipython
appnope==0.1.0; sys_platform == 'darwin' and platform_python_implementation != 'PyPy'
backports.shutil_get_terminal_size==1.0.0; platform_python_implementation != 'PyPy' and python_version == '2.7'
colorama==0.3.7; sys_platform == 'win32' and platform_python_implementation != 'PyPy'
pathlib2==2.1.0; platform_python_implementation != 'PyPy' and (python_version == '2.7' or python_version == '3.3')
pexpect==4.2.1; sys_platform != 'win32' and platform_python_implementation != 'PyPy'
pickleshare==0.7.4; platform_python_implementation != 'PyPy'
prompt_toolkit==1.0.9; platform_python_implementation != 'PyPy'
Pygments==2.1.3; platform_python_implementation != 'PyPy'
simplegeneric==0.8.1; platform_python_implementation != 'PyPy'
traitlets==4.3.1; platform_python_implementation != 'PyPy'
win_unicode_console==0.5; sys_platform == 'win32' and platform_python_implementation != 'PyPy' and python_version < '3.6'

# ipdb
ipython==5.1.0; platform_python_implementation != 'PyPy'

# pytest-cov -> cov-core
coverage==4.2

# pytest-cov -> pytest, detox -> tox
py==1.4.31

# pytest-cov
cov-core==1.15.0
pytest==3.0.5

# pytest-xdist
execnet==1.4.1

# And now the direct dependencies

# For better debugging
ipdb==0.10.1; platform_python_implementation != 'PyPy'

# Show log output for test failures
pytest-catchlog==1.2.2

# For code coverage statistics generation
pytest-cov==2.4.0

# Django integration for test runner
pytest-django==3.1.2

# Parallel test execution support
pytest-xdist==1.15.0

# For testing the Yappi profiler backend
yappi==0.98; platform_python_implementation == 'CPython' and python_version < '3.5'
